+++ msspackages/__init__.py
@@ -17,3 +17,5 @@ from .update_cwd import update_cwd_to_root
 from .data_ingestion import Pyspark_data_ingestion
 from .data_ingestion import find_multilevel_schema_items
 from .data_ingestion import pyspark_ingestion
+
+from .project_inital_setup import setup_runner