+++ msspackages/project_inital_setup/__init__.py
@@ -1,0 +1,1 @@
+from .understanding_eks_setup import setup_runner