+++ msspackages/project_inital_setup/console_setup_run.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+
+sudo yum install java
+
+python3 -m pip install -r requirements.txt
+
+
+#how to execute the script
+#bash console_setup_run.sh