+++ msspackages/project_inital_setup/notebook_setup_run.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
+
+
 apt update
 
 apt install sudo